Tests/LibWeb/Text/expected/abspos-relayout-toplayer-fallback.txt

open=true
dialog=70,40,80,50
partial-layouts=0 full-layouts=1
PASS
